% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFPredict-package.R
\docType{package}
\name{GFPredict-package}
\alias{GFPredict}
\alias{GFPredict-package}
\title{GFPredict: Predicts functionally related genes based on shared epigenomic and transcription factor binding patterns}
\description{
This package contains function: 'predicted_related_genes' that predicts coding and non-coding genes related to (user input) gene-sets. GFPredict compares the similarity between the epigenomic and transcription factor binding patterns at the promoter sites of the input genes and other genes, to make predictions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/reggenlab/GFPredict}
}

}
\author{
\strong{Maintainer}: Omkar Chandra \email{omkarc@iiitd.ac.in}

Authors:
\itemize{
  \item Vibhor Kumar \email{vibhor@iiitd.ac.in}
}

}
\keyword{internal}
