% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_related_genes.R
\name{predict_related_genes}
\alias{predict_related_genes}
\title{The function makes gene prediction based on shared transcription factors
and epigenomic marks present across promoter sites of the genes}
\usage{
predict_related_genes(genes, ml_model, n_bootstrap)
}
\arguments{
\item{genes}{Object of class list, containing functionally related genes}

\item{ml_model}{Object of class string, Machine learning model to select. Options: random.forest (default),
'xgboost' , 'svm', 'random.forest', 'linear.regression', 'logistic.regression'}

\item{n_bootstrap}{Number of bootstrapping of the features (n = 3 by default)}
}
\value{
List containing predicted gene, top predictors (if ml_model = 'random forest') and ML model performance metrics
}
\description{
The function makes gene prediction based on shared transcription factors
and epigenomic marks present across promoter sites of the genes
}
\examples{
breast_cancer_related_genes <- c("PTGS2", "ARID1A", "NFKB1", "TFF2", "STK11", "HRAS", "MSH2", "CTNNB1",
     "MEN1", "HIF1A", "MTHFR", "MAP2K4", "AKT1", "XRCC1", "S100P", "KLF5", "PARK2","NR5A2", "CLPTM1L", "GLI1",
     "TERT", "BRCA1", "SHH",	"PRSS1")

result <- predict_related_genes(breast_cancer_related_genes)

result <- predict_related_genes(breast_cancer_related_genes, model = 'xgboost', n_bootstrap = 5)

result[[1]] # list of predicted genes

result[[2]] # Table performance metrics 

#' result[[3]] # Table containing top predictors (if ml_model = random.forest)
Authors@R: person("Omkar Chandra", "Vibhor Kumar", email = "omkar@iiitd.ac.in",
role = c("Creator & author", "Author")
}
