% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_related_genes.R
\name{predict_related_genes}
\alias{predict_related_genes}
\title{The function makes gene prediction based on shared transcription factors
and epigenomic marks
Default machine learning model is "random.forest"}
\usage{
predict_related_genes(genes, ml_model, n_bootstrap)
}
\value{
list containing predicted gene, top predictors (if ml_model = 'random forest') and ML model metrics
}
\description{
The function makes gene prediction based on shared transcription factors
and epigenomic marks
Default machine learning model is "random.forest"
}
